move mux handler into server package
